app/requirements.txt
fastapi
uvicorn
pandas
openai
python-multipart
openpyxl
python-dotenv
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
PyJWT==1.7.0
requests
httpx
cryptography
clerk
svix
langgraph
langchain
langchain-community